Restore module length limit with focused cleanup
