+++ .rubocop.yml
@@ -74,7 +74,7 @@ Metrics/MethodLength:
   Max: 41
 
 Metrics/ModuleLength:
-  Max: 200
+  Max: 180
 
 RSpec/DescribeClass:
   Enabled: false
