+++ react_on_rails_pro/lib/react_on_rails_pro.rb
@@ -9,6 +9,7 @@ require "react_on_rails_pro/constants"
 require "react_on_rails_pro/compression_middleware_guard"
 require "react_on_rails_pro/engine"
 require "react_on_rails_pro/error"
+require "react_on_rails_pro/renderer_cache_path"
 require "react_on_rails_pro/utils"
 require "react_on_rails_pro/configuration"
 require "react_on_rails_pro/license_public_key"
