+++ react_on_rails_pro/spec/dummy/spec/pre_seed_renderer_cache_spec.rb
@@ -41,7 +41,7 @@ describe ReactOnRailsPro::PreSeedRendererCache do # rubocop:disable RSpec/FilePa
     # Clear env vars and deprecation warning guard
     ENV.delete("RENDERER_SERVER_BUNDLE_CACHE_PATH")
     ENV.delete("RENDERER_BUNDLE_PATH")
-    ReactOnRailsPro::Utils.send(:reset_renderer_bundle_path_deprecation_warned!)
+    ReactOnRailsPro::RendererCachePath.send(:reset_deprecation_warned!)
   end
 
   after do
