+++ react_on_rails_pro/spec/react_on_rails_pro/utils_spec.rb
@@ -499,71 +499,10 @@ module ReactOnRailsPro
     end
 
     describe ".resolve_renderer_cache_dir" do
-      before do
-        described_class.send(:reset_renderer_bundle_path_deprecation_warned!)
-      end
+      it "delegates to RendererCachePath for compatibility" do
+        allow(ReactOnRailsPro::RendererCachePath).to receive(:resolve).and_return("/tmp/renderer-cache")
 
-      after do
-        ENV.delete("RENDERER_SERVER_BUNDLE_CACHE_PATH")
-        ENV.delete("RENDERER_BUNDLE_PATH")
-        described_class.send(:reset_renderer_bundle_path_deprecation_warned!)
-      end
-
-      it "returns RENDERER_SERVER_BUNDLE_CACHE_PATH verbatim, including surrounding whitespace" do
-        ENV["RENDERER_SERVER_BUNDLE_CACHE_PATH"] = "  /tmp/renderer-cache  "
-
-        # Whitespace is preserved so the Rails pre-seed lands in the same path
-        # the Node renderer reads (it consumes the env var verbatim in
-        # configBuilder.ts).
-        expected_warning = Regexp.escape(
-          "RENDERER_SERVER_BUNDLE_CACHE_PATH has surrounding whitespace and will be used verbatim: " \
-          '"  /tmp/renderer-cache  "'
-        )
-        expect { @result = described_class.resolve_renderer_cache_dir }
-          .to output(/#{expected_warning}/).to_stderr
-        expect(@result).to eq("  /tmp/renderer-cache  ")
-      end
-
-      it "does not warn when RENDERER_SERVER_BUNDLE_CACHE_PATH has no surrounding whitespace" do
-        ENV["RENDERER_SERVER_BUNDLE_CACHE_PATH"] = "/tmp/renderer-cache"
-
-        expect { @result = described_class.resolve_renderer_cache_dir }.not_to output.to_stderr
-        expect(@result).to eq("/tmp/renderer-cache")
-      end
-
-      it "returns RENDERER_BUNDLE_PATH verbatim when only the deprecated var is set" do
-        ENV["RENDERER_BUNDLE_PATH"] = " /tmp/legacy-cache "
-
-        # Lookahead-based match so the assertion is order-independent: both the
-        # deprecation notice and the whitespace warning must appear, but either
-        # may print first since they come from different code paths.
-        whitespace_warning = Regexp.escape(
-          'RENDERER_BUNDLE_PATH has surrounding whitespace and will be used verbatim: " /tmp/legacy-cache "'
-        )
-        expected_warning = Regexp.new(
-          "(?=.*RENDERER_BUNDLE_PATH is deprecated)" \
-          "(?=.*#{whitespace_warning})",
-          Regexp::MULTILINE
-        )
-        expect { @result = described_class.resolve_renderer_cache_dir }
-          .to output(expected_warning)
-          .to_stderr
-        expect(@result).to eq(" /tmp/legacy-cache ")
-      end
-
-      it "raises when the preferred env var is whitespace-only" do
-        ENV["RENDERER_SERVER_BUNDLE_CACHE_PATH"] = "  "
-        ENV["RENDERER_BUNDLE_PATH"] = "/tmp/legacy-cache"
-
-        expect { described_class.resolve_renderer_cache_dir }
-          .to raise_error(ReactOnRailsPro::Error, /RENDERER_SERVER_BUNDLE_CACHE_PATH is whitespace-only/)
-      end
-
-      it "raises when the deprecated env var is whitespace-only" do
-        ENV["RENDERER_BUNDLE_PATH"] = "\t\n"
-
-        expect { described_class.resolve_renderer_cache_dir }
-          .to raise_error(ReactOnRailsPro::Error, /RENDERER_BUNDLE_PATH is whitespace-only/)
+        expect(described_class.resolve_renderer_cache_dir).to eq("/tmp/renderer-cache")
       end
     end
   end
